added socketcan compatibility, minor fixes, suggest: use python-can for serial too
