+++ 02_canSniffer_GUI/save/decodedPackets.csv
@@ -1,7 +1,3 @@
-,00000124,00,00,08,00,01,02,03,04,05,06,07
-,00000124,00,01,08,00,01,02,03,04,05,06,07
-,,,,,,,,,,,,
-prova,00000123 (test),00,01,08,00,01,02,03,04,05,06,07
 before ignition,621 (Immo),00,00,08,00,20,00,00,00,00,00,00
 Immo ignition,621 (Immo),00,01,08,00,48,50,00,00,00,00,00
 trunk open long press,160 (Remote),00,00,04,01,0C,7D,28,,,,
@@ -108,5 +104,3 @@ front left window stop,259,00,00,08,04,AE,02,08,00,00,00,00
 front left window down,259,00,00,08,04,AE,02,08,10,00,00,00
 front left window up,259,00,00,08,04,AE,02,08,08,00,00,00
 cabin light on,252,00,00,08,07,AE,03,80,00,00,00,00
-,00000124,00,00,08,00,01,02,03,04,05,06,07
-,00000124,00,01,08,01,01,02,03,04,05,06,07
